[fix] RecyclerViewの上からTextViewを置く形に変更
