+++ android/app/src/androidTest/java/com/wsr/infra/passwordgroup/queryservice/LocalGetAllPasswordGroupUseCaseQueryServiceTest.kt
@@ -24,7 +24,7 @@ import kotlin.test.Test
 
 @RunWith(AndroidJUnit4::class)
 @OptIn(ExperimentalCoroutinesApi::class)
-class RoomGetAllPasswordGroupUseCaseQueryServiceTest {
+class LocalGetAllPasswordGroupUseCaseQueryServiceTest {
     private lateinit var passwordGroupEntityDao: PasswordGroupEntityDao
     private lateinit var db: PassonDatabase
     private lateinit var target: LocalGetAllPasswordGroupUseCaseQueryServiceImpl
