+++ android/app/src/androidTest/java/com/wsr/infra/passwordgroup/repository/LocalPasswordGroupRepositoryCreateTest.kt
@@ -9,8 +9,8 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
 import com.wsr.email.Email
 import com.wsr.infra.PassonDatabase
-import com.wsr.infra.passwordgroup.PasswordGroupEntityDao
 import com.wsr.infra.passwordgroup.LocalPasswordGroupRepositoryImpl
+import com.wsr.infra.passwordgroup.PasswordGroupEntityDao
 import com.wsr.passwordgroup.PasswordGroup
 import com.wsr.passwordgroup.PasswordGroupId
 import com.wsr.passwordgroup.Remark
@@ -24,7 +24,7 @@ import kotlin.test.Test
 
 @RunWith(AndroidJUnit4::class)
 @OptIn(ExperimentalCoroutinesApi::class)
-class RoomPasswordGroupRepositoryCreateTest {
+class LocalPasswordGroupRepositoryCreateTest {
     private lateinit var passwordGroupEntityDao: PasswordGroupEntityDao
     private lateinit var db: PassonDatabase
     private lateinit var target: LocalPasswordGroupRepositoryImpl
