+++ android/app/src/androidTest/java/com/wsr/infra/passwordgroup/repository/LocalPasswordGroupRepositoryDeleteTest.kt
@@ -6,11 +6,11 @@ import android.content.Context
 import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.google.common.truth.Truth.assertThat
 import com.wsr.email.Email
+import com.wsr.exceptions.GetDataFailedException
 import com.wsr.infra.PassonDatabase
-import com.wsr.infra.passwordgroup.PasswordGroupEntityDao
 import com.wsr.infra.passwordgroup.LocalPasswordGroupRepositoryImpl
+import com.wsr.infra.passwordgroup.PasswordGroupEntityDao
 import com.wsr.passwordgroup.PasswordGroup
 import com.wsr.passwordgroup.PasswordGroupId
 import com.wsr.passwordgroup.Remark
@@ -21,10 +21,11 @@ import org.junit.runner.RunWith
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 import kotlin.test.Test
+import kotlin.test.assertFailsWith
 
 @RunWith(AndroidJUnit4::class)
 @OptIn(ExperimentalCoroutinesApi::class)
-class RoomPasswordGroupRepositoryUpdateTest {
+class LocalPasswordGroupRepositoryDeleteTest {
     private lateinit var passwordGroupEntityDao: PasswordGroupEntityDao
     private lateinit var db: PassonDatabase
     private lateinit var target: LocalPasswordGroupRepositoryImpl
@@ -43,9 +44,9 @@ class RoomPasswordGroupRepositoryUpdateTest {
         db.close()
     }
 
-    /*** update関数 ***/
+    /*** delete関数 ***/
     @Test
-    fun 新しいPasswordGroupの情報を渡すと指定されたPasswordGroupの更新を行う() = runTest {
+    fun passwordGroupIdを渡すと対応するPasswordGroupを削除する() = runTest {
         val mockedPasswordGroupId = PasswordGroupId("mockedPasswordGroupId")
         val mockedPasswordGroup = PasswordGroup(
             id = mockedPasswordGroupId,
@@ -55,17 +56,10 @@ class RoomPasswordGroupRepositoryUpdateTest {
         )
         target.create(mockedPasswordGroup)
 
-        val updatedMockedPasswordGroup = mockedPasswordGroup.copy(
-            title = Title("updatedMockedTitle"),
-            remark = Remark("updatedMockedRemark"),
-        )
-        target.update(
-            id = updatedMockedPasswordGroup.id,
-            title = updatedMockedPasswordGroup.title.value,
-            remark = updatedMockedPasswordGroup.remark.value,
-        )
+        target.delete(mockedPasswordGroupId)
 
-        val actual = passwordGroupEntityDao.getById(mockedPasswordGroupId.value)
-        assertThat(actual).isEqualTo(updatedMockedPasswordGroup)
+        assertFailsWith<GetDataFailedException.NoSuchElementException> {
+            passwordGroupEntityDao.getById(mockedPasswordGroupId.value)
+        }
     }
 }
