+++ android/app/src/androidTest/java/com/wsr/infra/passwordgroup/repository/LocalPasswordGroupRepositoryUpdateTest.kt
@@ -1,36 +1,33 @@
 @file:Suppress("NonAsciiCharacters", "TestFunctionName")
 
-package com.wsr.infra.passwordgroup.queryservice
+package com.wsr.infra.passwordgroup.repository
 
 import android.content.Context
 import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.google.common.truth.Truth
+import com.google.common.truth.Truth.assertThat
 import com.wsr.email.Email
-import com.wsr.exceptions.GetDataFailedException
 import com.wsr.infra.PassonDatabase
+import com.wsr.infra.passwordgroup.LocalPasswordGroupRepositoryImpl
 import com.wsr.infra.passwordgroup.PasswordGroupEntityDao
-import com.wsr.infra.passwordgroup.toEntity
 import com.wsr.passwordgroup.PasswordGroup
 import com.wsr.passwordgroup.PasswordGroupId
 import com.wsr.passwordgroup.Remark
 import com.wsr.passwordgroup.Title
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
-import org.junit.Test
 import org.junit.runner.RunWith
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
-import kotlin.test.assertFailsWith
+import kotlin.test.Test
 
 @RunWith(AndroidJUnit4::class)
 @OptIn(ExperimentalCoroutinesApi::class)
-class RoomGetPasswordGroupUseCaseQueryServiceTest {
-
+class LocalPasswordGroupRepositoryUpdateTest {
     private lateinit var passwordGroupEntityDao: PasswordGroupEntityDao
     private lateinit var db: PassonDatabase
-    private lateinit var target: LocalGetPasswordGroupUseCaseQueryServiceImpl
+    private lateinit var target: LocalPasswordGroupRepositoryImpl
 
     @BeforeTest
     fun setup() {
@@ -38,7 +35,7 @@ class RoomGetPasswordGroupUseCaseQueryServiceTest {
         db = Room.inMemoryDatabaseBuilder(context, PassonDatabase::class.java).build()
         passwordGroupEntityDao = db.passwordGroupEntityDao()
 
-        target = LocalGetPasswordGroupUseCaseQueryServiceImpl(passwordGroupEntityDao)
+        target = LocalPasswordGroupRepositoryImpl(passwordGroupEntityDao)
     }
 
     @AfterTest
@@ -46,9 +43,9 @@ class RoomGetPasswordGroupUseCaseQueryServiceTest {
         db.close()
     }
 
-    /*** getById関数 ***/
+    /*** update関数 ***/
     @Test
-    fun passwordGroupIdを渡せば対応するPasswordGroupを返す() = runTest {
+    fun 新しいPasswordGroupの情報を渡すと指定されたPasswordGroupの更新を行う() = runTest {
         val mockedPasswordGroupId = PasswordGroupId("mockedPasswordGroupId")
         val mockedPasswordGroup = PasswordGroup(
             id = mockedPasswordGroupId,
@@ -56,26 +53,19 @@ class RoomGetPasswordGroupUseCaseQueryServiceTest {
             title = Title("mockedTitle"),
             remark = Remark("mockedRemark"),
         )
-        passwordGroupEntityDao.insert(mockedPasswordGroup.toEntity())
+        target.create(mockedPasswordGroup)
 
-        val notTargetMockedPasswordGroup = PasswordGroup(
-            id = PasswordGroupId("notTargetPasswordGroupId"),
-            email = Email("notTargetMockedEmail"),
-            title = Title("notTargetMockedTitle"),
-            remark = Remark("notTargetMockedRemark"),
+        val updatedMockedPasswordGroup = mockedPasswordGroup.copy(
+            title = Title("updatedMockedTitle"),
+            remark = Remark("updatedMockedRemark"),
         )
-        passwordGroupEntityDao.insert(notTargetMockedPasswordGroup.toEntity())
+        target.update(
+            id = updatedMockedPasswordGroup.id,
+            title = updatedMockedPasswordGroup.title.value,
+            remark = updatedMockedPasswordGroup.remark.value,
+        )
 
-        val actual = target.getById(mockedPasswordGroupId)
-        Truth.assertThat(actual).isEqualTo(mockedPasswordGroup)
-    }
-
-    @Test
-    fun 存在しないpasswordGroupIdを渡せばNoSuchElementExceptionが投げられる() = runTest {
-        val mockedPasswordGroupId = PasswordGroupId("mockedPasswordGroupId")
-
-        assertFailsWith<GetDataFailedException.NoSuchElementException> {
-            target.getById(mockedPasswordGroupId)
-        }
+        val actual = passwordGroupEntityDao.getById(mockedPasswordGroupId.value)
+        assertThat(actual).isEqualTo(updatedMockedPasswordGroup)
     }
 }
